rename cfmm_common to common
